requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
